+++ CHANGELOG.md
@@ -16,6 +16,7 @@ v999 (May XX, 2021)
 * Improve project pages appearance: decrease action button width and left align text; widen from 9 to 10 columns main content.
 * Remove "Refresh Documents" button on task finished page because caches are now automatically cleared and document content refreshed.
 * Display system component component_state and component_type when component is listed for a system.
+* Add simple "back" link to question to take user to previous question.
 
 **Developer changes**